Move collection.dart to schema
